Refactoring: removed duplicated interfaces
