+++ src/hooks/usePlatforms.ts
@@ -1,9 +1,9 @@
 import { useQuery } from "@tanstack/react-query";
 import apiClient from "../services/api-client";
 import { FetchResponse } from "../services/api-client";
-import { Platform } from "./useGames";
 
-interface Plaforms {
+
+export interface Platform {
     id: number;
     name: string;
     slug: string;
